refactor: 나의 문제집에서 폴더 구조 컴포넌트 분리
