+++ src/pages/Library.tsx
@@ -16,6 +16,7 @@ import Spinner from '@/shared/components/Spinner';
 import RightClickMenu from '@/features/library/components/RightClickMenu/RightClickMenu';
 import RightClickMenuItem from '@/features/library/components/RightClickMenu/RightClickMenuItem';
 import RightClickMenuDivider from '@/features/library/components/RightClickMenu/RightClickMenuDivider';
+import FolderList from '@/shared/components/FolderList';
 
 interface Folder {
   id: number;
@@ -25,7 +26,6 @@ interface Folder {
 }
 
 const QUESTION_SET_TYPE = 'QUESTION_SET';
-const ALL_FOLDER_ID = 1;
 
 const Container = styled.div`
   display: flex;
@@ -57,109 +57,6 @@ const FileListSearchInput = styled.input`
   &:focus {
     outline: none;
     border-color: ${({ theme }) => theme.colors.semantic.primary};
-  }
-`;
-
-const FolderContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-bottom: 16px;
-`;
-
-const FolderTag = styled.div<{
-  isDragOver?: boolean;
-  folderColor: string;
-  folderHoverColor: string;
-  isActive?: boolean;
-}>`
-  display: inline-flex;
-  align-items: center;
-  padding: 6px 12px;
-  border-radius: ${({ theme }) => theme.radius.radius2};
-  background-color: ${({ isDragOver, folderColor, folderHoverColor, isActive }) =>
-    isDragOver || isActive ? folderHoverColor : folderColor};
-  border: 1px solid
-    ${({ isDragOver, folderHoverColor, isActive }) =>
-      isDragOver || isActive ? folderHoverColor : 'transparent'};
-  font-size: ${({ theme }) => theme.typography.body3Regular.fontSize};
-  color: white;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  user-select: none;
-  font-weight: ${({ isActive }) => (isActive ? '700' : '500')};
-  box-shadow: ${({ isActive }) => (isActive ? '0 2px 8px rgba(0, 0, 0, 0.15)' : 'none')};
-
-  &:hover {
-    background-color: ${({ folderHoverColor }) => folderHoverColor};
-    border-color: ${({ folderHoverColor }) => folderHoverColor};
-    color: white;
-  }
-`;
-
-const AddFolderButton = styled.div`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  padding: 6px 12px;
-  border-radius: ${({ theme }) => theme.radius.radius2};
-  background-color: ${({ theme }) => theme.colors.background.foreground};
-  border: 1px dashed ${({ theme }) => theme.colors.border.border1};
-  font-size: ${({ theme }) => theme.typography.body3Regular.fontSize};
-  color: ${({ theme }) => theme.colors.text.default};
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  user-select: none;
-
-  &:hover {
-    background-color: ${({ theme }) => theme.colors.semantic.primary};
-    border-color: ${({ theme }) => theme.colors.semantic.primary};
-    border-style: solid;
-    color: white;
-  }
-`;
-
-const FolderInputContainer = styled.div`
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-  padding: 6px 12px;
-  border-radius: ${({ theme }) => theme.radius.radius2};
-  background-color: ${({ theme }) => theme.colors.background.foreground};
-  border: 1px solid ${({ theme }) => theme.colors.semantic.primary};
-`;
-
-const FolderInput = styled.input`
-  border: none;
-  outline: none;
-  background: transparent;
-  font-size: ${({ theme }) => theme.typography.body3Regular.fontSize};
-  color: ${({ theme }) => theme.colors.text.default};
-  width: 120px;
-  padding: 0;
-
-  &::placeholder {
-    color: ${({ theme }) => theme.colors.text.subtitle};
-  }
-`;
-
-const FolderActionButton = styled.button`
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 14px;
-  transition: transform 0.1s;
-
-  &:hover {
-    transform: scale(1.2);
-  }
-
-  &:active {
-    transform: scale(0.9);
   }
 `;
 
@@ -348,47 +245,21 @@ interface QuestionSetApiResponse {
   questionSets: QuestionSets;
 }
 
-const FOLDER_COLORS = [
-  { bg: '#3b82f6', hover: '#2563eb' }, // 파란색
-  { bg: '#10b981', hover: '#059669' }, // 초록색
-  { bg: '#8b5cf6', hover: '#7c3aed' }, // 보라색
-  { bg: '#f59e0b', hover: '#d97706' }, // 주황색
-  { bg: '#ec4899', hover: '#db2777' }, // 분홍색
-  { bg: '#06b6d4', hover: '#0891b2' }, // 청록색
-  { bg: '#ef4444', hover: '#dc2626' }, // 빨간색
-  { bg: '#6366f1', hover: '#4f46e5' }, // 인디고
-  { bg: '#14b8a6', hover: '#0d9488' }, // 틸
-  { bg: '#f97316', hover: '#ea580c' }, // 오렌지
-];
-
-const getFolderColor = (folderId: number) => {
-  const index = folderId % FOLDER_COLORS.length;
-  return FOLDER_COLORS[index];
-};
-
 const Library = () => {
   const [searchTerm, setSearchTerm] = useState('');
   const [debouncedSearchTerm, setDebouncedSearchTerm] = useState('');
   const [editingItemId, setEditingItemId] = useState<number | null>(null);
   const [editingTitle, setEditingTitle] = useState('');
   const [isVisibleMenu, setIsVisibleMenu] = useState<boolean>(false);
-  const [selectedCell, setSelectedCell] = useState<QuestionSetContentType | null>(null);
-  const [draggedItem, setDraggedItem] = useState<QuestionSetContentType | null>(null);
-  const [dragOverFolderId, setDragOverFolderId] = useState<number | null>(null);
-  const [isAddingFolder, setIsAddingFolder] = useState(false);
-  const [newFolderName, setNewFolderName] = useState('');
+  const [selectedCell, setSelectedCell] = useState<QuestionSetContentType | null>(
+    null,
+  );
+  const [draggedItem, setDraggedItem] = useState<QuestionSetContentType | null>(
+    null,
+  );
   const [selectedFolderId, setSelectedFolderId] = useState<number | null>(null);
-  const [isVisibleFolderMenu, setIsVisibleFolderMenu] = useState<boolean>(false);
-  const [selectedFolder, setSelectedFolder] = useState<Folder | null>(null);
-  const [editingFolderId, setEditingFolderId] = useState<number | null>(null);
-  const [editingFolderName, setEditingFolderName] = useState('');
 
   const [mousePoint, setMousePoint] = useState<{
-    x: number;
-    y: number;
-  }>({ x: 0, y: 0 });
-
-  const [folderMousePoint, setFolderMousePoint] = useState<{
     x: number;
     y: number;
   }>({ x: 0, y: 0 });
@@ -407,18 +278,6 @@ const Library = () => {
     setMousePoint({ x: e.clientX, y: e.clientY });
   };
 
-  const handleFolderContextMenu = (
-    e: React.MouseEvent<HTMLDivElement, MouseEvent>,
-    folder: Folder,
-  ) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    setSelectedFolder(folder);
-    setIsVisibleFolderMenu(true);
-    setFolderMousePoint({ x: e.clientX, y: e.clientY });
-  };
-
   const updateTitleMutation = useMutation({
     mutationFn: ({ id, title }: { id: number | undefined; title: string }) => {
       if (id === undefined || id === null) {
@@ -430,7 +289,9 @@ const Library = () => {
       return api.patch(`/question-set/${id}`, { title });
     },
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['questionSets', selectedFolderId] });
+      queryClient.invalidateQueries({
+        queryKey: ['questionSets', selectedFolderId],
+      });
       setEditingItemId(null);
       setEditingTitle('');
     },
@@ -444,7 +305,9 @@ const Library = () => {
       return api.delete(`/question-set/${id}`);
     },
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['questionSets', selectedFolderId] });
+      queryClient.invalidateQueries({
+        queryKey: ['questionSets', selectedFolderId],
+      });
     },
     onError: (error) => {
       alert(`삭제 중 에러가 발생했습니다: ${error.message}`);
@@ -452,7 +315,13 @@ const Library = () => {
   });
 
   const moveFolderMutation = useMutation({
-    mutationFn: ({ questionSetId, folderId }: { questionSetId: number; folderId: number }) => {
+    mutationFn: ({
+      questionSetId,
+      folderId,
+    }: {
+      questionSetId: number;
+      folderId: number;
+    }) => {
       return api.patch(`/question-set/${questionSetId}`, {
         commonFolderId: folderId,
       });
@@ -463,56 +332,6 @@ const Library = () => {
     },
     onError: (error) => {
       alert(`폴더 이동 중 에러가 발생했습니다: ${error.message}`);
-    },
-  });
-
-  const createFolderMutation = useMutation({
-    mutationFn: (name: string) => {
-      return api.post('/common-folders', {
-        name,
-        type: QUESTION_SET_TYPE,
-      });
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['folders'] });
-      setIsAddingFolder(false);
-      setNewFolderName('');
-    },
-    onError: (error) => {
-      alert(`폴더 생성 중 에러가 발생했습니다: ${error.message}`);
-    },
-  });
-
-  const deleteFolderMutation = useMutation({
-    mutationFn: (id: number) => {
-      return api.delete(`/common-folders/${id}`);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['folders'] });
-      queryClient.invalidateQueries({ queryKey: ['questionSets'] });
-      if (selectedFolderId === selectedFolder?.id) {
-        setSelectedFolderId(null);
-      }
-    },
-    onError: (error) => {
-      alert(`폴더 삭제 중 에러가 발생했습니다: ${error.message}`);
-    },
-  });
-
-  const updateFolderNameMutation = useMutation({
-    mutationFn: ({ id, name }: { id: number; name: string }) => {
-      if (!name.trim()) {
-        throw new Error('폴더 이름은 비워둘 수 없습니다.');
-      }
-      return api.patch(`/common-folders/${id}`, { name, type: QUESTION_SET_TYPE });
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['folders'] });
-      setEditingFolderId(null);
-      setEditingFolderName('');
-    },
-    onError: (error) => {
-      alert(error.message);
     },
   });
 
@@ -560,12 +379,13 @@ const Library = () => {
   const { data: folders, isPending: isFoldersPending } = useQuery({
     queryKey: ['folders'],
     queryFn: async () => {
-      const res = await api.get<Folder[]>(`/common-folders?type=${QUESTION_SET_TYPE}`);
+      const res = await api.get<Folder[]>(
+        `/common-folders?type=${QUESTION_SET_TYPE}`,
+      );
       return res.data.sort((a, b) => a.sortOrder - b.sortOrder);
     },
   });
 
-  // 폴더가 로드되면 첫 번째 폴더를 자동 선택
   useEffect(() => {
     if (folders && folders.length > 0 && selectedFolderId === null) {
       setSelectedFolderId(folders[0].id);
@@ -588,64 +408,23 @@ const Library = () => {
     },
     enabled: selectedFolderId !== null,
     refetchInterval: (query) =>
-      query.state.data?.questionSets.content.some((item) => item.status === 'PENDING')
+      query.state.data?.questionSets.content.some(
+        (item) => item.status === 'PENDING',
+      )
         ? 5000
         : false,
   });
 
-  const handleDragStart = (e: React.DragEvent<HTMLDivElement>, item: QuestionSetContentType) => {
+  const handleDragStart = (
+    e: React.DragEvent<HTMLDivElement>,
+    item: QuestionSetContentType,
+  ) => {
     setDraggedItem(item);
     e.dataTransfer.effectAllowed = 'move';
   };
 
   const handleDragEnd = () => {
     setDraggedItem(null);
-    setDragOverFolderId(null);
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>, folderId: number) => {
-    e.preventDefault();
-    e.dataTransfer.dropEffect = 'move';
-    setDragOverFolderId(folderId);
-  };
-
-  const handleDragLeave = () => {
-    setDragOverFolderId(null);
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>, folder: Folder) => {
-    e.preventDefault();
-    setDragOverFolderId(null);
-
-    if (!draggedItem) return;
-
-    moveFolderMutation.mutate({
-      questionSetId: draggedItem.questionSetId,
-      folderId: folder.id,
-    });
-
-    setDraggedItem(null);
-  };
-
-  const handleAddFolder = () => {
-    setIsAddingFolder(true);
-  };
-
-  const handleCancelAddFolder = () => {
-    setIsAddingFolder(false);
-    setNewFolderName('');
-  };
-
-  const handleConfirmAddFolder = () => {
-    if (!newFolderName.trim()) {
-      alert('폴더 이름을 입력해주세요.');
-      return;
-    }
-    createFolderMutation.mutate(newFolderName.trim());
-  };
-
-  const handleFolderClick = (folderId: number) => {
-    setSelectedFolderId(folderId);
   };
 
   const handleMenuRename = useCallback(() => {
@@ -678,44 +457,6 @@ const Library = () => {
     [selectedCell, moveFolderMutation],
   );
 
-  const handleFolderMenuRename = useCallback(() => {
-    if (!selectedFolder) return;
-    setEditingFolderId(selectedFolder.id);
-    setEditingFolderName(selectedFolder.name);
-    setIsVisibleFolderMenu(false);
-  }, [selectedFolder]);
-
-  const handleFolderMenuDelete = useCallback(async () => {
-    if (!selectedFolder) return;
-    setIsVisibleFolderMenu(false);
-
-    try {
-      const response = await api.get<{ questionSetCount: number }>(
-        `/common-folders/${selectedFolder.id}/delete-warning`,
-      );
-
-      const questionSetCount = response.data.questionSetCount;
-      const confirmMessage =
-        questionSetCount > 0
-          ? `'${selectedFolder.name}' 폴더에 ${questionSetCount}개의 문제집이 있습니다.\n정말로 삭제하시겠습니까?`
-          : `'${selectedFolder.name}' 폴더를 정말 삭제하시겠습니까?`;
-
-      if (window.confirm(confirmMessage)) {
-        deleteFolderMutation.mutate(selectedFolder.id);
-      }
-    } catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : '알 수 없는 오류';
-      alert(`폴더 삭제 경고 정보를 가져오는데 실패했습니다: ${errorMessage}`);
-    }
-  }, [selectedFolder, deleteFolderMutation]);
-
-  const submitFolderNameEdit = (folder: Folder) => {
-    updateFolderNameMutation.mutate({
-      id: folder.id,
-      name: editingFolderName,
-    });
-  };
-
   if (isPending || isFoldersPending) {
     return <Spinner />;
   }
@@ -731,7 +472,11 @@ const Library = () => {
 
   return (
     <Container>
-      <RightClickMenu isVisible={isVisibleMenu} setIsVisible={setIsVisibleMenu} point={mousePoint}>
+      <RightClickMenu
+        isVisible={isVisibleMenu}
+        setIsVisible={setIsVisibleMenu}
+        point={mousePoint}
+      >
         <RightClickMenuItem
           icon="✏️"
           title="문제집 이름 변경"
@@ -777,119 +522,27 @@ const Library = () => {
           disabled={selectedCell?.status !== 'COMPLETE'}
         />
       </RightClickMenu>
-      <RightClickMenu
-        isVisible={isVisibleFolderMenu}
-        setIsVisible={setIsVisibleFolderMenu}
-        point={folderMousePoint}
-      >
-        <RightClickMenuItem
-          icon="✏️"
-          title="폴더 이름 변경"
-          onClick={handleFolderMenuRename}
-          disabled={selectedFolder?.id === ALL_FOLDER_ID}
-        />
-        <RightClickMenuItem
-          icon="❌"
-          title="폴더 삭제"
-          onClick={handleFolderMenuDelete}
-          disabled={selectedFolder?.id === ALL_FOLDER_ID}
-        />
-      </RightClickMenu>
+
       <LibraryWrapper>
         <LibraryTitle />
         <LibraryProgressSummary percent={data?.learningProgress ?? 0} />
         <Spacer height="12px" />
-        {/* 검색 input창 -> 디바운싱 구현되어 있습니다. */}
         <FileListSearchInput
           placeholder="문제집 제목으로 검색"
           value={searchTerm}
           onChange={(e) => setSearchTerm(e.target.value)}
         />
         <Spacer height="12px" />
-        <FolderContainer>
-          {folders &&
-            folders.map((folder) => {
-              const colors = getFolderColor(folder.id);
-              const isEditingThisFolder = editingFolderId === folder.id;
-              return (
-                <FolderTag
-                  key={folder.id}
-                  isDragOver={dragOverFolderId === folder.id}
-                  isActive={selectedFolderId === folder.id}
-                  folderColor={colors.bg}
-                  folderHoverColor={colors.hover}
-                  onClick={() => handleFolderClick(folder.id)}
-                  onContextMenu={(e) => handleFolderContextMenu(e, folder)}
-                  onDragOver={(e) => handleDragOver(e, folder.id)}
-                  onDragLeave={handleDragLeave}
-                  onDrop={(e) => handleDrop(e, folder)}
-                >
-                  {isEditingThisFolder ? (
-                    <>
-                      📁{' '}
-                      <FolderInput
-                        value={editingFolderName}
-                        onChange={(e) => setEditingFolderName(e.target.value)}
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') {
-                            submitFolderNameEdit(folder);
-                          }
-                          if (e.key === 'Escape') {
-                            setEditingFolderId(null);
-                          }
-                        }}
-                        onClick={(e) => e.stopPropagation()}
-                        autoFocus
-                      />
-                      <FolderActionButton
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          submitFolderNameEdit(folder);
-                        }}
-                      >
-                        ✔️
-                      </FolderActionButton>
-                      <FolderActionButton
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          setEditingFolderId(null);
-                        }}
-                      >
-                        ❌
-                      </FolderActionButton>
-                    </>
-                  ) : (
-                    <>📁 {folder.name}</>
-                  )}
-                </FolderTag>
-              );
-            })}
-          {isAddingFolder ? (
-            <FolderInputContainer>
-              <span>📁</span>
-              <FolderInput
-                value={newFolderName}
-                onChange={(e) => setNewFolderName(e.target.value)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    handleConfirmAddFolder();
-                  }
-                  if (e.key === 'Escape') {
-                    handleCancelAddFolder();
-                  }
-                }}
-                placeholder="폴더 이름"
-                autoFocus
-              />
-              <FolderActionButton onClick={handleConfirmAddFolder}>✔️</FolderActionButton>
-              <FolderActionButton onClick={handleCancelAddFolder}>❌</FolderActionButton>
-            </FolderInputContainer>
-          ) : (
-            <AddFolderButton onClick={handleAddFolder}>➕</AddFolderButton>
-          )}
-        </FolderContainer>
+        <FolderList
+          folders={folders}
+          selectedFolderId={selectedFolderId}
+          onFolderSelect={setSelectedFolderId}
+          draggedItem={draggedItem}
+          onItemDrop={(folderId, questionSetId) => {
+            moveFolderMutation.mutate({ questionSetId, folderId });
+          }}
+        />
         <Spacer height="12px" />
-        {/* 여기에서 부터 리스트 박스입니다.*/}
         <ListBox>
           <ListRow>
             <HeaderCell align="left">문제집</HeaderCell>
@@ -901,7 +554,11 @@ const Library = () => {
           </ListRow>
 
           {[...filteredQuestionSets]
-            .sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()) // 시간 내림차순
+            .sort(
+              (a, b) =>
+                new Date(b.createdAt).getTime() -
+                new Date(a.createdAt).getTime(),
+            )
             .map((item) => {
               const isEditing = editingItemId === item.questionSetId;
 
@@ -931,8 +588,12 @@ const Library = () => {
                           autoFocus
                         />
                         <div>
-                          <EditIconButton onClick={() => submitTitleEdit(item)}>✔️</EditIconButton>
-                          <EditIconButton onClick={() => setEditingItemId(null)}>❌</EditIconButton>
+                          <EditIconButton onClick={() => submitTitleEdit(item)}>
+                            ✔️
+                          </EditIconButton>
+                          <EditIconButton onClick={() => setEditingItemId(null)}>
+                            ❌
+                          </EditIconButton>
                         </div>
                       </TitleContainer>
                     ) : (
@@ -943,7 +604,9 @@ const Library = () => {
                   </ListCell>
                   <ListCell>{item.questionCount}</ListCell>
                   <ListCell>
-                    {new Intl.DateTimeFormat('sv-SE').format(new Date(item.createdAt))}
+                    {new Intl.DateTimeFormat('sv-SE').format(
+                      new Date(item.createdAt),
+                    )}
                   </ListCell>
                   <ListCell>{TYPE_MAP[item.questionType] ?? '생성 실패'}</ListCell>
                   <StatusCell status={item.status}>
@@ -951,7 +614,9 @@ const Library = () => {
                   </StatusCell>
                   <ListCell>
                     {item.status === 'COMPLETE' && (
-                      <PrimaryButton onClick={() => handleSolveClick(item.questionSetId)}>
+                      <PrimaryButton
+                        onClick={() => handleSolveClick(item.questionSetId)}
+                      >
                         풀기
                       </PrimaryButton>
                     )}
